# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/auth/wecom/callback:
    get:
      parameters:
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 状态码
        in: query
        name: state
        required: true
        type: string
      produces:
      - text/html
      responses: {}
      summary: 企业微信授权回调
      tags:
      - 认证
...
swagger: "2.0"